Strip pathname portion off all message routines that print filename:line.
Print additional messages show the elapsed time and transfer rate for the spool portion of a spooled backup job and also for the tape write portion.

Add DriveType directive to the Director's Include Option FileSet resource.  Allowed values are: fixed, removable, cdrom, and remote.  There is only an implementation for Windows because it is the only platform that has the concept of drives.

Adds EnhancedWild directive to the Director's Include Option FileSet resource.  Allowed values are: yes and no.

	When EnhancedWild is enabled then the processing of the Wild, WildDir and WildFile is changed in the following ways.

	Patterns conform to Posix
		\ is not a special character in character classification []
			To match a - it must be the first or last character
			To match a ] it must be the first character

	fnmatch option FNM_FILE_NAME is specified
		* doesn't match a / so it won't match multiple directory levels in a path

	Relative WildFile patterns (ones without a leading /) match against the filename portion
		This in combination with the FNM_FILE_NAME fnmatch() flag makes directives such as WildFile = abc*.def work as expected.

	Adds support for the shell's feature of brace expansion.


git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@3178 91ce42f0-d328-0410-95d8-f526ca767f89