Reworked the MinGW32  build-dependencies script, it has been renamed to build-depkgs-mingw32.  It now handles two options -h to display help and -C to reextract the source and apply the patches.  The -C option is useful for updating the source after new patches are released.
Moved the 3rd party package URLs to a separate file External-mingw32.  This allows us to update to new versions without having to modify the script and makes it easier to figure out which packages and which versions are used.

Added a new script build-depkgs-msvc.cmd which does the same thing as build-depkgs-mingw32 except they are built using Microsoft Visual C and placed in the depkgs-msvc directory.  There is also an External-msvc file to describe the packages.

Updated README.mingw32 with the updated instructions.

Added a couple of missing files to the installer for PostgreSQL.


git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@3282 91ce42f0-d328-0410-95d8-f526ca767f89