crc32 optimization
This is a port of the Linux kernel crc32 function I developed
over the years. This impl. is several times faster than the one it
replaces. Adapted to bacula usage.

The new crc32 function needs to know the endian of the
computing CPU and it uses HAVE_BIG_ENDIAN and HAVE_LITTLE_ENDIAN
defines to deduce what endian the CPU uses.