Fix bug #1751 Make duplicate checking somewhat less naiv.
We were almost there with excluding duplicate checks for Copy and
Migration jobs. We only used the wrong variable. Now allocated a
new one and made the naming somewhat more comprehensible. And extra
check is added that jobs inserted by Migration and Copy and which
have the flag set are also not compared against normal jobs in the
normal canceling process so you should be able to run a copy job
now and start the normal backup job without it being canceled as
reported by alltech in the bug report.